Bulk renaming
Open for advice, this is one of my first codes/programs and feedback would be fantastic.

This script asks for the path you want to work on and make sure if there are sub or nested folders you don't want being effected to remove before hand and then it asking for the keyword of selection for the renaming then it asks for the new name/the part that it will rename with.